fix: use cases and tests
