+++ src/modules/user/infra/exceptions/AlreadyUsedToken.exception/index.ts
@@ -1,0 +1,7 @@
+import { HttpException, HttpStatus } from '@nestjs/common';
+
+export class AlreadyUsedTokenException extends HttpException {
+  constructor() {
+    super('Esse token já foi utilizado.', HttpStatus.BAD_REQUEST);
+  }
+}
