+++ src/modules/user/infra/exceptions/EmailAlreadyInUse.exception/index.ts
