+++ src/modules/user/infra/exceptions/InvalidToken.exception/index.ts
@@ -1,0 +1,7 @@
+import { HttpException, HttpStatus } from '@nestjs/common';
+
+export class InvalidTokenException extends HttpException {
+  constructor() {
+    super('Token inválido ou expirado.', HttpStatus.BAD_REQUEST);
+  }
+}
