+++ src/modules/user/infra/usecases/InviteProfessor.usecase/dto.ts
@@ -1,0 +1,8 @@
+import { InviteResponseDTO } from '../../requests/InviteProfessor.request/dto';
+import { InviteBodyDTO } from '../../requests/InviteProfessor.request/dto';
+
+export interface InviteProfessorUseCaseDTO extends InviteBodyDTO {
+  user_email: string;
+}
+
+export interface InviteProfessorUseCaseResponseDTO extends InviteResponseDTO {}
