+++ src/modules/user/infra/usecases/ValidateToken.usecase/dto.ts
@@ -1,0 +1,9 @@
+import {
+  ValidateTokenBodyDTO,
+  ValidateTokenResponseDTO,
+} from 'src/modules/user/infra/requests/ValidateToken.request/dto';
+
+export interface ValidateTokenUseCaseDTO extends ValidateTokenBodyDTO {}
+
+export interface ValidateTokenUseCaseResponseDTO
+  extends ValidateTokenResponseDTO {}
